lua-luajit: remove duplicated lua-jit and merge to lua-luajit.
